requirements.txt
aiofiles==24.1.0
anyio==4.10.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Deprecated==1.2.18
get-annotations==0.1.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-tools==1.71.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.0
packaging==25.0
pillow==11.3.0
protobuf==5.29.5
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
python-docx==1.2.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
Telethon==1.40.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
wrapt==1.17.2
yandex-cloud-ml-sdk==0.13.1
yandexcloud==0.354.0
